perf: improve the FillMapWithStrAndError performance and split too long test case
