+++ pkg/util/mapz/map.go
@@ -1,7 +1,7 @@
 package mapz
 
 func FillMapWithStrAndError(keys []string, value error) map[string]error {
-	retMap := make(map[string]error)
+	retMap := make(map[string]error, len(keys))
 	if len(keys) == 0 {
 		return retMap
 	}
